Calibrate RDU motor temperature sensor
Data comes from sandwiching a thermocouple between the motor sensor and
a large lump of mild steel. The mild steel was heated externally with a
heat gun to 100 degrees then allowed to passively cool. Finally ice was
placed on the steel to get down to 0 degrees. The motor sensor
resistance and thermocouple temperature were recorded during the entire
process.